+++ R/autoplot.R
@@ -16,7 +16,31 @@
 #' @param X A matrix-like R object (e.g., a data frame or matrix) containing 
 #' ONLY the feature columns from the training data.
 #' 
-#' @param ... Additional optional arguments. (Currently ignored.)
+#' @param color_by Character string specifying an optional feature column in 
+#' \code{X} to use for coloring whenever \code{type = "dependence"}.
+#' 
+#' @param smooth Logical indicating whether or not to add a smoother to the
+#' scatterplot whenever \code{type = "dependence"}. Default is \code{TRUE}.
+#' 
+#' @param smooth_color The color to use for the smoother whenever 
+#' \code{smooth = TRUE}. The default is \code{"black"}; see 
+#' \code{\link[ggplot2]{geom_smooth}} for details.
+#' 
+#' @param smooth_linetype The type of line to use for the smoother whenever 
+#' \code{smooth = TRUE}. The default is \code{"solid"}; see 
+#' \code{\link[ggplot2]{geom_smooth}} for details.
+#' 
+#' @param smooth_size The size to use for the smoother whenever 
+#' \code{smooth = TRUE}. The default is \code{1}; see 
+#' \code{\link[ggplot2]{geom_smooth}} for details.
+#' 
+#' @param smooth_alpha The transparency to use for the smoother whenever 
+#' \code{smooth = TRUE}. The default is \code{1}; see 
+#' \code{\link[ggplot2]{geom_smooth}} for details.
+#' 
+#' @param ... Additional optional arguments to be passed onto 
+#' \code{\link[ggplot2]{geom_col}} (if \code{type = "importance"}) or 
+#' \code{\link[ggplot2]{geom_point}} (if \code{type = "dependence"}).
 #' 
 #' @importFrom ggplot2 aes_string autoplot coord_flip geom_col geom_point
 #' 
@@ -26,19 +50,30 @@
 #'
 #' @export
 autoplot.fastshap <- function(object, type = c("importance", "dependence"),
-                              feature = NULL, X, ...) {
+                              feature = NULL, X, color_by = NULL, 
+                              smooth = FALSE, smooth_color = "red", 
+                              smooth_linetype = "solid", smooth_size = 1, 
+                              smooth_alpha = 1, ...) {
   type <- match.arg(type)
   if (type == "importance") {
+    
+    # Construct data to plot
     shap_imp <- data.frame(
       Variable = names(object),
-      Importance = apply(object, MARGIN = 2, FUN = function(x) sum(abs(x)))
+      Importance = apply(object, MARGIN = 2, FUN = function(x) mean(abs(x)))
     )
-    ggplot(shap_imp, aes_string("reorder(Variable, Importance)", "Importance")) +
-      geom_col() +
+    
+    # Construct plot
+    x_string <- "reorder(Variable, Importance)"
+    p <- ggplot(shap_imp, aes_string(x_string, "Importance")) +
+      geom_col(...) +
       coord_flip() +
       xlab("") +
       ylab("mean(|Shapley value|)")
+    
   } else {
+    
+    # Construct data to plot
     if (is.null(feature)) {
       feature <- names(object)[1L]
     }
@@ -46,10 +81,28 @@ autoplot.fastshap <- function(object, type = c("importance", "dependence"),
       x = X[, feature, drop = TRUE], 
       y = object[, feature, drop = TRUE]
     )
-    ggplot(shap_dep, aes_string(x = "x", y = "y")) +
-      geom_point(alpha = 0.3) +
-      geom_smooth() +
+    if (!is.null(color_by)) {
+      shap_dep[[color_by]] <- X[[color_by]]
+    }
+
+    # Construct plot
+    p <- if (!is.null(color_by)) {
+      ggplot(shap_dep, aes_string(x = "x", y = "y", color = color_by))
+    } else {
+      ggplot(shap_dep, aes_string(x = "x", y = "y"))
+    }
+    p <- p +
+      geom_point(...) +
       xlab(feature) +
       ylab("Shapley value")
+    if (smooth) {
+      p <- p + geom_smooth(se = FALSE, color = smooth_color,
+                           linetype = smooth_linetype, size = smooth_size,
+                           alpha = smooth_alpha)
+    }
   }
+  
+  # Return plot
+  p
+
 }